made handleControllererrors util

+++ backend/src/controllers/PublicationController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express';
 import Publication from '../models/Publication';
+import { handleControllerErrors } from '../utils/handleControllerErrors';
 
 export const getPublication = async (
     req: Request,
@@ -7,9 +8,11 @@ export const getPublication = async (
 ): Promise<Response> => {
     try {
         const publication = await Publication.find({});
-        return res.status(200).json(publication);
-    } catch (error: any) {
-        return res.status(500).json({ message: error.message });
+        if (!publication) throw new Error("Couldn't find any publications");
+
+        return res.status(200).json({data: {found: true, publication}});
+    } catch (err) {
+        return handleControllerErrors(err, res, 'Publications not found');
     }
 };
 
@@ -20,9 +23,19 @@ export const getPublicationById = async (
     try {
         const { id } = req.params;
         const publicationById = await Publication.findById(id);
-        return res.status(200).json(publicationById);
-    } catch (error: any) {
-        return res.status(500).json({ message: error.message });
+
+        if (!id)
+            throw new Error('Publication id not found. Wrong request params');
+        if (!publicationById)
+            throw new Error('Could not find this publication');
+
+        return res.status(200).json({data: {found: true, publicationById}});
+    } catch (err) {
+        return handleControllerErrors(
+            err,
+            res,
+            'Could not find this publication',
+        );
     }
 };
 
@@ -32,9 +45,17 @@ export const createPublication = async (
 ): Promise<Response> => {
     try {
         const newPublication = await Publication.create(req.body);
-        return res.status(200).json(newPublication);
-    } catch (error: any) {
-        return res.status(500).json({ message: error.message });
+
+        if (!newPublication)
+            throw new Error('Publication could not be created. Wrong params');
+
+        return res.status(200).json({data: {created: true, newPublication}});
+    } catch (err) {
+        return handleControllerErrors(
+            err,
+            res,
+            'Publication could not be created',
+        );
     }
 };
 
@@ -44,14 +65,21 @@ export const updatePublication = async (
 ): Promise<Response> => {
     try {
         const { id } = req.params;
+        if (!id)
+            throw new Error(
+                'Could not find publication id for update. Wrong id',
+            );
+
         const updatedPublication = await Publication.findByIdAndUpdate(
             id,
             req.body,
         );
+        if (!updatedPublication)
+            throw new Error('Could not update publication. Wrong params;');
 
-        return res.status(200).json(updatedPublication);
-    } catch (error: any) {
-        return res.status(500).json({ message: error.message });
+        return res.status(200).json({data: {updated: true, updatedPublication}});
+    } catch (err) {
+        return handleControllerErrors(err, res, 'Could not update publication');
     }
 };
 
@@ -61,12 +89,13 @@ export const deletePublication = async (
 ): Promise<Response> => {
     try {
         const { id } = req.params;
-        const deletedPublication = await Publication.findByIdAndDelete(id);
+        if (!id) throw new Error("Could not delete publication. Wrong id");
 
-        return res.status(200).json(deletedPublication);
-    } catch (error: any) {
-        return res.status(500).json({ message: error.message });
+        const deletedPublication = await Publication.findByIdAndDelete(id);
+        if (!deletedPublication) throw new Error("Could not delete publication.");
+
+        return res.status(200).json({data: {deleted: true, deletedPublication}});
+    } catch (err) {
+        return handleControllerErrors(err, res, "Could not delete publication.");
     }
 };
-
-//TODO make an error / unfound handler
